+++ tasks/seq/gordeva_t_max_val_of_column_matrix/src/ops_seq.cpp
@@ -47,9 +47,9 @@ bool gordeva_t_max_val_of_column_matrix_seq::TestTaskSequential::run() {
   // int cols = input_[0].size();
   // int rows = input_.size();
 
-  for (int i = 0; i < input_[0].size(); i++) {
+  for (size_t i = 0; i < input_[0].size(); i++) {
     int max_el = input_[0][i];
-    for (int j = 1; j < input_.size(); j++)
+    for (size_t j = 1; j < input_.size(); j++)
       if (input_[j][i] > max_el) max_el = input_[j][i];
 
     res_[i] = max_el;
@@ -63,7 +63,7 @@ bool gordeva_t_max_val_of_column_matrix_seq::TestTaskSequential::post_processing
 
   int* output_matr = reinterpret_cast<int*>(taskData->outputs[0]);
 
-  for (int i = 0; i < res_.size(); i++) output_matr[i] = res_[i];
+  for (size_t i = 0; i < res_.size(); i++) output_matr[i] = res_[i];
   return true;
 }
